ssl/requirements.txt
kivy>=2.3.0
pygame>=2.6.0
pillow>=10.0.0
